refactor: :recycle: Simplifica a configuração do sistema introduzindo funções para configurar o CORS, criar a instância do aplicativo, habilitar pipes globais e habilitar o controle de versão. Valida os dados do usuário com função dedicada.
