+++ .husky/pre-commit
@@ -1,3 +1,3 @@
-pnpm lint
+pnpm lint --fix
 pnpm test
 pnpm build
