+++ src/middlewares/index.ts
@@ -1,3 +1,2 @@
-export * from "./auth.middleware";
 export * from "./formatResponse.middleware";
 export * from "./logger.middleware";
