+++ src/modules/logger/index.ts
@@ -1,0 +1,2 @@
+export * from "./logger.module";
+export * from "./logger.service";
