+++ src/modules/logger/logger.module.ts
@@ -1,0 +1,9 @@
+
+import { Module } from '@nestjs/common';
+import { CustomLogger } from './logger.service';
+
+@Module({
+  providers: [CustomLogger],
+  exports: [CustomLogger],
+})
+export class LoggerModule {}
