+++ src/modules/sector/sector.controller.ts
@@ -1,5 +1,5 @@
 import { SectorWithoutIdDto, UpdateSectorDto } from "@/typing";
-import { IsPublic } from "@/utils";
+import { IsPublic } from "@/utils/decorators";
 import {
     Body,
     Controller,
