+++ src/utils/decorators/UniqueRegister.decorator.ts
@@ -1,0 +1,14 @@
+import { registerDecorator, ValidationOptions } from "class-validator";
+import { UniqueRegisterValidator } from "../validators";
+
+export const UniqueRegister = (validationOptions?: ValidationOptions) => {
+    return (object: object, property: string) => {
+      registerDecorator({
+        target: object.constructor,
+        propertyName: property,
+        options: validationOptions,
+        constraints: [],
+        validator: UniqueRegisterValidator,
+      });
+    };
+  };
