+++ src/utils/decorators/index.ts
@@ -1,1 +1,3 @@
 export * from "./IsPublic.decorator";
+export * from "./UniqueEmail.decorator";
+export * from "./UniqueRegister.decorator";
