+++ src/utils/functions/index.ts
@@ -5,3 +5,5 @@ export * from "./encryptPassword";
 export * from "./getAuthInformation";
 export * from "./getUserDataByToken";
 export * from "./searchUserByRegister";
+export * from "./systemConfiguration";
+export * from "./validUserData";
