--- src/utils/guards/role.guard.ts
@@ -1,14 +1,0 @@
-import { CanActivate, ExecutionContext, Injectable } from "@nestjs/common";
-import { Observable } from "rxjs";
-
-@Injectable()
-export class RoleGuard implements CanActivate {
-  canActivate(
-    context: ExecutionContext,
-  ): boolean | Promise<boolean> | Observable<boolean> {
-    console.log({
-      roleSystem: context,
-    });
-    return true;
-  }
-}
