+++ src/utils/index.ts
@@ -2,9 +2,9 @@ export * from "./alias";
 export * from "./constants";
 export * from "./filters";
 export * from "./functions";
-export * from "./guards";
+export * from "../guards";
 export * from "./mocks";
 export * from "./validators";
-export * from "./middlewares";
+export * from "../middlewares";
 export * from "./decorators";
 // export * from "./resources";
